CLOUDSYNC-1: Ability to copy Managed Datastreams
Implements includeManaged option for Copy tasks. When selected (the default),
objects that include managed datastreams are copied along with the content of
those datastreams. When a DuraCloud store is the destination, the contentId
of each datastream is that of the FOXML, with "+dsId+dsVersionId" appended.
When a Fedora store is the destination, the managed content is first staged
via the /upload endpoint, and the references within the FOXML are modified
as appropriate prior to ingest.

https://jira.duraspace.org/browse/CLOUDSYNC-1