CLOUDSYNC-9: Ability to copy E/R datastreams
Implements two new boolean options for Copy tasks, one for E and one for R
datastreams. If checked, that type of datastream, when encountered at the source,
will be converted to a managed datastream in transit to the destination. The
default behavior (when the option is unchecked) is to ignore the datastream
content and keep the original reference in place.

https://jira.duraspace.org/browse/CLOUDSYNC-9